test(16): Refactor test packages to better match src directory
