+++ src/test/java/com/locust/OperatorStarterTests.java
@@ -1,6 +1,5 @@
-package com.locust.operator;
+package com.locust;
 
-import com.locust.LocustTestOperatorStarter;
 import com.locust.operator.controller.LocustTestReconciler;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.server.mock.EnableKubernetesMockClient;
@@ -11,7 +10,7 @@ import org.junit.jupiter.api.TestInstance;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
-import static com.locust.operator.controller.utils.TestFixtures.executeWithK8sMockServer;
+import static com.locust.operator.controller.TestFixtures.executeWithK8sMockServer;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @EnableKubernetesMockClient(https = false, crud = true)
