+++ src/test/java/com/locust/operator/controller/utils/resource/manage/ResourceCreationManagerTests.java
@@ -1,8 +1,6 @@
-package com.locust.operator.controller;
+package com.locust.operator.controller.utils.resource.manage;
 
 import com.locust.operator.controller.utils.LoadGenHelpers;
-import com.locust.operator.controller.utils.resource.manage.ResourceCreationHelpers;
-import com.locust.operator.controller.utils.resource.manage.ResourceCreationManager;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.server.mock.EnableKubernetesMockClient;
 import lombok.extern.slf4j.Slf4j;
@@ -15,11 +13,11 @@ import org.junit.jupiter.api.TestInstance;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
+import static com.locust.operator.controller.TestFixtures.assertK8sResourceCreation;
+import static com.locust.operator.controller.TestFixtures.containerEnvironmentMap;
+import static com.locust.operator.controller.TestFixtures.executeWithK8sMockServer;
+import static com.locust.operator.controller.TestFixtures.prepareNodeConfig;
 import static com.locust.operator.controller.dto.OperationalMode.MASTER;
-import static com.locust.operator.controller.utils.TestFixtures.assertK8sResourceCreation;
-import static com.locust.operator.controller.utils.TestFixtures.containerEnvironmentMap;
-import static com.locust.operator.controller.utils.TestFixtures.executeWithK8sMockServer;
-import static com.locust.operator.controller.utils.TestFixtures.prepareNodeConfig;
 import static org.mockito.Mockito.when;
 
 @Slf4j
